[gh-49]: move tests to __tests__ folder
