+++ .dockerignore
@@ -14,4 +14,4 @@ node_modules
 .vscode
 
 # Directory with tests
-test
+__tests__
