+++ __tests__/fixtures/bukovel-ticket-skipass-responce.html
