+++ package.json
@@ -11,7 +11,7 @@
     "build:server": "gulp build:server",
     "build:client": "gulp build:client",
     "lint": "standard --verbose | snazzy",
-    "test": "npm run lint && ava",
+    "test": "jest",
     "test:watch": "ava --watch"
   },
   "dependencies": {
@@ -85,9 +85,12 @@
   },
   "devDependencies": {
     "ava": "0.17.0",
+    "babel-jest": "18.0.0",
     "enzyme": "2.7.0",
+    "jest": "18.1.0",
     "pify": "2.3.0",
     "react-addons-test-utils": "15.4.2",
+    "react-test-renderer": "15.4.2",
     "sinon": "1.17.7",
     "snazzy": "5.0.0",
     "standard": "8.6.0"
